Refactoring index import
