+++ packages/commons/src/auth/authenticationMiddleware.ts
@@ -2,12 +2,9 @@
 import { ZodiosRouterContextRequestHandler } from "@zodios/express";
 import { makeApiProblemBuilder } from "pagopa-interop-models";
 import { match } from "ts-pattern";
-import {
-  ExpressContext,
-  fromAppContext,
-  JWTConfig,
-  jwtFromAuthHeader,
-} from "../index.js";
+import { ExpressContext, fromAppContext } from "../context/index.js";
+import { JWTConfig } from "../config/index.js";
+import { jwtFromAuthHeader } from "./headers.js";
 import { readAuthDataFromJwtToken, verifyJwtToken } from "./jwt.js";
 
 const makeApiProblem = makeApiProblemBuilder({});
