+++ packages/commons/src/eservice-documents/eserviceDocumentUtils.ts
@@ -20,7 +20,9 @@ import {
 } from "pagopa-interop-models";
 import { match, P } from "ts-pattern";
 import { z, ZodError } from "zod";
-import { calculateChecksum, FileManager, Logger } from "../index.js";
+import { calculateChecksum } from "../utils/fileUtils.js";
+import { FileManager } from "../file-manager/index.js";
+import { Logger } from "../logging/index.js";
 import {
   parseOpenApi,
   restApiFileToBuffer,
