+++ packages/commons/src/logging/loggerMiddleware.ts
@@ -2,7 +2,7 @@
 import * as express from "express";
 import { AppContext } from "../context/context.js";
 import { getUserInfoFromAuthData } from "../auth/authData.js";
-import { LoggerMetadata, logger } from "./index.js";
+import { LoggerMetadata, logger } from "../logging/index.js";
 
 export function loggerMiddleware(serviceName: string): express.RequestHandler {
   return (req, res, next): void => {
